Move the adapters into their own packages.
